requirements.txt
asgiref==3.6.0
Django==3.2.18
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-environ==0.10.0
django-uuidfield==0.5.0
djangorestframework==3.14.0
environ==1.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==0.21.1
pytz==2023.3
requests==2.28.1
sqlparse==0.4.4
whitenoise==6.4.0
numpy==1.21.6
pandas==1.5.2
